# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ALT Agency Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List all workspaces for current user
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{id}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    patch:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdate"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{id}/members:
    get:
      operationId: listWorkspaceMembers
      tags: [workspaces]
      summary: List workspace members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceMember"
    post:
      operationId: inviteWorkspaceMember
      tags: [workspaces]
      summary: Invite a member to workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInviteInput"
      responses:
        "201":
          description: Invite sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMember"

  # ── Audits ─────────────────────────────────────────────────────────────────
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List all proposals
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create a new proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Created proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get proposal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    patch:
      operationId: updateProposal
      tags: [proposals]
      summary: Update proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalUpdate"
      responses:
        "200":
          description: Updated proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    delete:
      operationId: deleteProposal
      tags: [proposals]
      summary: Delete a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /proposals/{id}/send:
    post:
      operationId: sendProposal
      tags: [proposals]
      summary: Send proposal to client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}/accept:
    post:
      operationId: acceptProposal
      tags: [proposals]
      summary: Accept a proposal (client action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
...
components:
  schemas:
...
    Workspace:
      type: object
      required: [id, name, slug, ownerId, plan, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        ownerId:
          type: integer
        plan:
          type: string
        logoUrl:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        customDomain:
          type: ["string", "null"]
        memberCount:
          type: integer
        auditCount:
          type: integer
        leadCount:
          type: integer
        createdAt:
          type: string
    WorkspaceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: string
    WorkspaceUpdate:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
        primaryColor:
          type: string
        customDomain:
          type: string
    WorkspaceMember:
      type: object
      required: [id, workspaceId, userId, role, status, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        userId:
          type: ["integer", "null"]
        email:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
        status:
          type: string
        createdAt:
          type: string
    MemberInviteInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
        role:
          type: string

    # ── Audit ─────────────────────────────────────────────────────────────────
...
    Proposal:
      type: object
      required: [id, workspaceId, title, status, totalValue, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        clientId:
          type: ["integer", "null"]
        leadId:
          type: ["integer", "null"]
        title:
          type: string
        status:
          type: string
        totalValue:
          type: number
        currency:
          type: string
        validUntil:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        acceptedAt:
          type: ["string", "null"]
        viewCount:
          type: integer
        services:
          type: array
          items:
            $ref: "#/components/schemas/ProposalService"
        createdAt:
          type: string
    ProposalService:
      type: object
      required: [id, name, price]
      properties:
        id:
          type: integer
        proposalId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        quantity:
          type: integer
        recurring:
          type: boolean
    ProposalInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        clientId:
          type: integer
        leadId:
          type: integer
        currency:
          type: string
        validUntil:
          type: string
        services:
          type: array
          items:
            $ref: "#/components/schemas/ProposalServiceInput"
    ProposalServiceInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: integer
        recurring:
          type: boolean
    ProposalUpdate:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        currency:
          type: string
        validUntil:
          type: string
...